requirements.txt
# SmartLink Dependencies
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
flask-wtf>=1.2.2
flask-mail>=0.10.0
sqlalchemy>=2.0.41
werkzeug>=3.1.3
wtforms>=3.2.1
dnspython>=2.7.0
requests>=2.32.4

# Development dependencies
python-dotenv>=1.0.0